requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pyyaml>=6.0

# cuTAGI - Bayesian Neural Network library
# Install from: https://github.com/lhnguyen102/cuTAGI
pytagi>=0.2.1

# Data loading
torch>=1.10.0
torchvision>=0.11.0

# Testing
pytest>=7.0.0

# Development
tqdm>=4.62.0
pandas>=1.3.0
